refactor: move api auth & provider
